final commit

+++ backend/controllers/goalController.js
@@ -6,7 +6,7 @@ const Goal=require('../model/goalModel')
 //@access private
 
 const getGoals=asyncHandler(async(req,res)=>{
-    let goal= await Goal.find({})
+    let goal= await Goal.find({user:req.user.id})
     res.status(200).json(goal)
 })
 
@@ -21,7 +21,8 @@ const setGoal=asyncHandler(async(req,res)=>{
 
 
     let goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
 
 
@@ -44,6 +45,13 @@ const updateGoal=asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
        throw  new Error("item not found")
+
+
+    }
+
+    if(goal.user.toString()!==req.user.id){
+        res.status(401)
+        throw new Error('Access Denaied.')
     }
 
     let updatedGoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{
@@ -66,6 +74,10 @@ const deleteGoal=asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error ("item not found")
+    }
+    if(goal.user.toString()!==req.user.id){
+        res.status(401)
+        throw new Error('Access Denaied.')
     }
     await goal.deleteOne()
